{dkrajzew} merging with trunk
git-svn-id: http://svn.code.sf.net/p/sumo/code/branches/dkrajzew@13656 29f149b2-0c63-11e1-8f5c-a347f6a9acd0

+++ src/microsim/trigger/MSCalibrator.cpp
@@ -3,7 +3,7 @@
 /// @author  Daniel Krajzewicz
 /// @author  Jakob Erdmann
 /// @date    Tue, May 2005
-/// @version $Id: MSCalibrator.cpp 2345 2013-01-17 14:27:26Z behr_mi $
+/// @version $Id$
 ///
 // Calibrates the flow on an edge by removing an inserting vehicles
 /****************************************************************************/
@@ -437,7 +437,7 @@ MSCalibrator::remainingVehicleCapacity(int laneIndex) const {
         // ensure vehicles can be inserted on short edges
         return MAX2(1, (int)(myEdge->getLength() / spacePerVehicle));
     } else {
-        return last->getPositionOnLane() / spacePerVehicle;
+        return (int)(last->getPositionOnLane() / spacePerVehicle);
     }
 }
